requirements.txt
Django==1.10
django-bootstrap3==8.0.0
django-taggit==0.22.1
django-timezone-field==2.0
olefile==0.44
Pillow==4.2.1
pytz==2017.2
